feat: now settings wil be restored.
so after exiting the app it will save the state to `savefile` file
then when opening the app it will read from the `savefile`.

NOTE: since it just raw byte dump it will crash if the savefile is
broken. So if the app wont open try deleting the `savefile` and on that
topic too if the `savefile` generated at `x86_64` cpu it will crash on
the arm cpu and the other way around.